fix: allow decimal input in adjusted hours, test cases, and actual hours fields; fix weighted value precision to match Excel; enable collaboration tab only when all component columns are filled; pre-load collaboration rows with zero defaults

+++ InitialEstimatePOC/Models/ProjectEntity.cs
@@ -93,11 +93,11 @@ public class ProjectEntity
 
     // === Test Cases for System Testing ===
     public bool UseTestCasesForEstimate { get; set; }
-    public int TestCasesSimple { get; set; }
-    public int TestCasesMedium { get; set; }
-    public int TestCasesComplex { get; set; }
-    public int TestCasesVeryComplex { get; set; }
-    public decimal TestCaseIterations { get; set; } = 1m;
+    public decimal TestCasesSimple { get; set; }
+    public decimal TestCasesMedium { get; set; }
+    public decimal TestCasesComplex { get; set; }
+    public decimal TestCasesVeryComplex { get; set; }
+    public decimal TestCaseIterations { get; set; } = 1;
 
     /// <summary>Oracle: CREATED_DATE DATE</summary>
     public DateTime CreatedDate { get; set; } = DateTime.UtcNow;
